Linux Install: Recommend not mounting <library_path> directly
Protect users from data loss if <library_path> unmounted and Navidrome scans the library by recommending to store <library_path> in a subdirectory of the mount point. Navidrome deletes all songs from the database if <library_path> exists but is an empty mount point, permanently deleting playlist contents, favourites and play counts. If <library_path> does not exist, the scan simply fails and the database is not modified.