Move security stuff back to the security page
The security considerations are already centralized in a single page, it should
not be changed for just one aspect of Navidrome.

I also changed the HPP vulnerability section to make it less scary, as this is
a pretty normal concern for reverse proxy authentication, and hopefully
navidrome already plays nice with this. It is also not very usefull to mention
hardening by stripping other credential sources without listing the sources
(and I don't get exactly how all of them work yet).