Clarify use-case for ReverseProxyWhitelist option
I have seen several people on reddit using this option with a
non-authenticating proxy setup, mostly when the proxy is only used to route
traffic or handle SSL termination, which can easily lead to an insecure setup
(e.g. the proxy might not remove by default the user header set by a malicious
client).

The option's name is a bit unfortunate, as it suggests that it should be used
for all proxy setups, while it is actually usefull only for authenticating
proxies.